+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Added
 - Add commands to query the word with modification to command palette.
 - Add menus to the editor context menu.
-- Refactoring: Add CscopeExecute, CscopeConfig and CscopeLog class.
+- Refactoring: Add CscopeExecute, CscopeConfig, CscopeLog and CscopeHistory class.
 - Refactoring: Split CscopePosition class and move some functionality.
 
 ## [0.0.4]